Refactor everything into new TherapySettingsViewModelDelegate
The number of function aliases in `TherapySettingsViewModel` was getting to be a mess. I consolidated `SaveCompletion`, `MaxTempBasalSavePreflight`, `SyncBasalRateSchedule`, and `SyncDeliveryLimits` all into a new `TherapySettingsViewModelDelegate`. Things got a bit cleaner, and now `DeviceDataManager` handles doing both the `maxTempBasalSavePreflight` and `syncDeliveryLimits` so it is not so loose of a contract.

This meant some gyrations were needed here in OnboardingViewModel.